+++ packages/uniforms-bootstrap5/src/ErrorField.tsx
@@ -1,0 +1,29 @@
+import classnames from 'classnames';
+import React, { HTMLProps } from 'react';
+import { connectField, filterDOMProps, Override } from 'uniforms';
+
+export type ErrorFieldProps = Override<
+  Omit<HTMLProps<HTMLDivElement>, 'onChange'>,
+  { error?: any; errorMessage?: string }
+>;
+
+function Error({
+  children,
+  className,
+  error,
+  errorMessage,
+  ...props
+}: ErrorFieldProps) {
+  return !error ? null : (
+    <div
+      className={classnames('card', 'mb-3', className)}
+      {...filterDOMProps(props)}
+    >
+      <div className="card-body">
+        {children ? children : <h4 className="card-title">{errorMessage}</h4>}
+      </div>
+    </div>
+  );
+}
+
+export default connectField(Error, { initialValue: false, kind: 'leaf' });