+++ packages/uniforms-bridge-json-schema/src/JSONSchemaBridge.ts
@@ -55,12 +55,12 @@ function extractValue(...xs: (boolean | null | string | undefined)[]) {
 }
 
 function pathToName(path: string) {
-  path = path.startsWith('.')
-    ? path
-        .replace(/\['(.+?)'\]/g, '.$1')
+  path = path.startsWith('/')
+    ? path.replace(/\//g, '.').replace(/~0/g, '~').replace(/~1/g, '/')
+    : path
+        .replace(/\[('|")(.+?)\1\]/g, '.$2')
         .replace(/\[(.+?)\]/g, '.$1')
-        .replace(/\\'/g, "'")
-    : path.replace(/\//g, '.').replace(/~0/g, '~').replace(/~1/g, '/');
+        .replace(/\\'/g, "'");
 
   return path.slice(1);
 }
@@ -81,9 +81,9 @@ export default class JSONSchemaBridge extends Bridge {
     this.schema = distinctSchema(schema);
 
     // Memoize for performance and referential equality.
-    this.getField = memoize(this.getField);
-    this.getSubfields = memoize(this.getSubfields);
-    this.getType = memoize(this.getType);
+    this.getField = memoize(this.getField.bind(this));
+    this.getSubfields = memoize(this.getSubfields.bind(this));
+    this.getType = memoize(this.getType.bind(this));
   }
 
   getError(name: string, error: any) {