+++ packages/uniforms-material/__tests__/LongTextField.tsx
@@ -97,7 +97,7 @@ test('<LongTextField> - renders a TextField which correctly reacts on change', (
   );
 
   expect(wrapper.find(TextField)).toHaveLength(1);
-  // @ts-ignore Provide a valid EventTarget.
+  // @ts-expect-error Provide a valid EventTarget.
   wrapper.find(TextField).props().onChange!({ target: { value: 'y' } });
   expect(onChange).toHaveBeenLastCalledWith('x', 'y');
 });
@@ -112,7 +112,7 @@ test('<LongTextField> - renders a TextField which correctly reacts on change (em
   );
 
   expect(wrapper.find(TextField)).toHaveLength(1);
-  // @ts-ignore Provide a valid EventTarget.
+  // @ts-expect-error Provide a valid EventTarget.
   wrapper.find(TextField).props().onChange!({ target: { value: '' } });
   expect(onChange).toHaveBeenLastCalledWith('x', '');
 });
@@ -127,7 +127,7 @@ test('<LongTextField> - renders a TextField which correctly reacts on change (sa
   );
 
   expect(wrapper.find(TextField)).toHaveLength(1);
-  // @ts-ignore Provide a valid EventTarget.
+  // @ts-expect-error Provide a valid EventTarget.
   wrapper.find(TextField).props().onChange!({ target: { value: 'y' } });
   expect(onChange).toHaveBeenLastCalledWith('x', 'y');
 });