+++ packages/uniforms-semantic/src/RadioField.tsx
@@ -15,7 +15,7 @@ export type RadioFieldProps = HTMLFieldProps<
   {
     allowedValues?: string[];
     checkboxes?: boolean;
-    transform?(value: string): string;
+    transform?: (value: string) => string;
   }
 >;