+++ scripts/setupEnzyme.ts
@@ -1,5 +1,5 @@
+import Adapter from '@wojtekmaj/enzyme-adapter-react-17';
 import Enzyme from 'enzyme';
-import Adapter from 'enzyme-adapter-react-16';
 
 Object.defineProperty(window, 'matchMedia', {
   writable: true,