+++ website/components/Playground.tsx
@@ -2,7 +2,7 @@ import ConfigProvider from 'antd/lib/config-provider';
 import classNames from 'classnames';
 import omit from 'lodash/omit';
 import React, { Component } from 'react';
-// @ts-ignore
+// @ts-expect-error
 import Frame, { FrameContextConsumer } from 'react-frame-component';
 import { ValidatedForm, connectField, context, useForm } from 'uniforms';
 
@@ -19,7 +19,7 @@ export class Playground extends Component<any, any> {
   }
 
   constructor() {
-    // @ts-ignore: Types.
+    // @ts-expect-error: Types.
     super(...arguments);
 
     const state = schema.clean(parseQuery());
@@ -62,10 +62,12 @@ export class Playground extends Component<any, any> {
 
   render() {
     return (
-      // @ts-ignore: Types.
+      // @ts-expect-error: Types.
       <PlaygroundForm
         className={playgroundStyles['playground']}
         model={this.state}
+        // It's bound in constructor.
+        // eslint-disable-next-line @typescript-eslint/unbound-method
         onChange={this.onChange}
         schema={bridge}
       >
@@ -121,7 +123,7 @@ const PlaygroundModelDebug = () => {
 
 class PlaygroundPreview extends Component<any, any> {
   constructor() {
-    // @ts-ignore: Types.
+    // @ts-expect-error: Types.
     super(...arguments);
 
     this._schema = eval(`(${this.props.value.schema})`);